Add support for ESM modules
This generates ESM module syntax ("export default")
in a separate esm/ directory.

For simplicity, it assumes that if you want to use
ESM modules, you also have ES6 features in your
environment, so the build script turns that on as well.

The esm index.js is exposed through the "module" field
in package.json.